ledger-api-common: Move the HealthService here.
