+++ ledger/ledger-api-common/BUILD.bazel
@@ -61,9 +61,11 @@ testDependencies = [
     "//daml-lf/language",
     "//daml-lf/parser",
     "//language-support/scala/bindings",
-    "//ledger-api/testing-utils",
+    "//ledger/ledger-api-akka",
     "//ledger-api/rs-grpc-bridge",
     "//ledger-api/rs-grpc-akka",
+    "//ledger-api/rs-grpc-testing-utils",
+    "//ledger-api/testing-utils",
     "//ledger/ledger-api-domain:ledger-api-domain",
     "//ledger/participant-state",
     "//daml-lf/archive:daml_lf_archive_reader",
@@ -73,6 +75,7 @@ testDependencies = [
     "@maven//:com_typesafe_akka_akka_actor_2_12",
     "@maven//:com_typesafe_akka_akka_stream_2_12",
     "@maven//:com_typesafe_akka_akka_stream_testkit_2_12",
+    "@maven//:com_typesafe_akka_akka_testkit_2_12",
     "@maven//:com_typesafe_scala_logging_scala_logging_2_12",
     "@maven//:org_reactivestreams_reactive_streams",
     "@maven//:org_scalactic_scalactic_2_12",
