+++ ledger/ledger-api-common/src/test/suite/scala/com/digitalasset/platform/server/api/services/grpc/GrpcHealthServiceSpec.scala
@@ -1,11 +1,11 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.sandbox.health
+package com.digitalasset.platform.server.api.services.grpc
 
 import com.digitalasset.grpc.adapter.server.rs.MockServerCallStreamObserver
 import com.digitalasset.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import com.digitalasset.platform.sandbox.health.HealthServiceSpec._
+import com.digitalasset.platform.server.api.services.grpc.GrpcHealthServiceSpec._
 import io.grpc.health.v1.health.{HealthCheckRequest, HealthCheckResponse}
 import org.scalatest.concurrent.Eventually
 import org.scalatest.time.{Second, Span}
@@ -13,13 +13,13 @@ import org.scalatest.{Matchers, WordSpec}
 
 import scala.concurrent.{Await, ExecutionContext}
 
-object HealthServiceSpec {
+object GrpcHealthServiceSpec {
   private val request = HealthCheckRequest()
 
   private val servingResponse = HealthCheckResponse(HealthCheckResponse.ServingStatus.SERVING)
 }
 
-final class HealthServiceSpec
+final class GrpcHealthServiceSpec
     extends WordSpec
     with Matchers
     with Eventually
@@ -32,7 +32,7 @@ final class HealthServiceSpec
 
   "HealthService" should {
     "check the current health" in {
-      val service = new HealthService()
+      val service = new GrpcHealthService()
 
       val response = Await.result(service.check(request), patienceConfig.timeout)
 
@@ -41,7 +41,7 @@ final class HealthServiceSpec
 
     "observe changes in health" in {
       val responseObserver = new MockServerCallStreamObserver[HealthCheckResponse]
-      val service = new HealthService()
+      val service = new GrpcHealthService()
 
       service.watch(request, responseObserver)
       responseObserver.demandResponse()
