+++ ledger/sandbox/BUILD.bazel
@@ -34,7 +34,6 @@ compileDependencies = [
     "//language-support/scala/bindings",
     "//ledger-api/rs-grpc-akka",
     "//ledger-api/rs-grpc-bridge",
-    "//ledger-api/rs-grpc-testing-utils",
     "//ledger-service/jwt",
     "//ledger/ledger-api-akka",
     "//ledger/ledger-api-auth",
@@ -176,7 +175,6 @@ testDependencies = [
     "//ledger/ledger-api-common:ledger-api-common-scala-tests-lib",
     "//ledger/test-common:test-common",
     "//ledger-api/testing-utils",
-    "@maven//:com_typesafe_akka_akka_testkit_2_12",
     "@maven//:commons_io_commons_io",
     "@maven//:org_awaitility_awaitility",
     "@maven//:org_mockito_mockito_core",
